requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.43
feedparser==6.0.11
requests==2.32.3
APScheduler==3.10.4
matplotlib==3.8.4
pyyaml==6.0.2
